add FD polarizability example, remove all other examples
