--- examples/.gitignore
@@ -1,1 +1,0 @@
-/*.pdf
