+++ examples/polarizability.jl
@@ -1,108 +1,54 @@
-# # Polarizability by linear response
+# # Polarizability using automatic differentiation
 #
-# We compute the polarizability of a Helium atom. The polarizability
-# is defined as the change in dipole moment
-# ```math
-# \mu = \int r ρ(r) dr
-# ```
-# with respect to a small uniform electric field ``E = -x``.
-#
-# We compute this in two ways: first by finite differences (applying a
-# finite electric field), then by linear response. Note that DFTK is
-# not really adapted to isolated atoms because it uses periodic
-# boundary conditions. Nevertheless we can simply embed the Helium
-# atom in a large enough box (although this is computationally wasteful).
-#
-# As in other tests, this is not fully converged, convergence
-# parameters were simply selected for fast execution on CI,
+# For a more classical approach and more details about computing polarizabilities,
+# see [Polarizability by linear response](@ref).
 
 using DFTK
 using LinearAlgebra
+using DFTK: FallbackFunctional, LibxcFunctional
 
-a = 10.
-lattice = a * I(3)  # cube of ``a`` bohrs
-## Helium at the center of the box
-atoms     = [ElementPsp(:He, psp=load_psp("hgh/lda/He-q2"))]
-positions = [[1/2, 1/2, 1/2]]
+## Construct PlaneWaveBasis given a particular electric field strength
+## Again we take the example of a Helium atom.
+function make_basis_model(ε::T; a=10., Ecut=30) where T
+    lattice=T(a) * I(3)  # lattice is a cube of ``a`` Bohrs
+    He = ElementPsp(:He, psp=load_psp("hgh/lda/He-q2"))
+    atoms = [He]
+    positions = [[1/2; 1/2; 1/2]] # Helium at the center of the box
 
-
-kgrid = [1, 1, 1]  # no k-point sampling for an isolated system
-Ecut = 30
-tol = 1e-8
+    terms = [
+        Kinetic(),
+        AtomicLocal(),
+        AtomicNonlocal(),
+        Ewald(),
+        PspCorrection(),
+        Hartree(),
+        ExternalFromReal(r -> -ε * (r[1] - a/2)),
+        Xc([FallbackFunctional{:lda}(:lda_x), FallbackFunctional{:lda}(:lda_c_vwn)])
+    ]
+    model = Model(lattice, atoms, positions; terms, symmetries=false)
+    PlaneWaveBasis(model; Ecut, kgrid=[1, 1, 1])  # No k-point sampling on isolated system
+end
 
 ## dipole moment of a given density (assuming the current geometry)
 function dipole(basis, ρ)
+    @assert isdiag(basis.model.lattice)
+    a  = basis.model.lattice[1, 1]
     rr = [a * (r[1] - 1/2) for r in r_vectors(basis)]
     sum(rr .* ρ) * basis.dvol
-end;
-
-# ## Using finite differences
-# We first compute the polarizability by finite differences.
-# First compute the dipole moment at rest:
-model = model_LDA(lattice, atoms, positions; symmetries=false)
-basis = PlaneWaveBasis(model; Ecut, kgrid)
-res   = self_consistent_field(basis, tol=tol)
-μref  = dipole(basis, res.ρ)
-
-# Then in a small uniform field:
-ε = .01
-model_ε = model_LDA(lattice, atoms, positions;
-                    extra_terms=[ExternalFromReal(r -> -ε * (r[1] - a/2))],
-                    symmetries=false)
-basis_ε = PlaneWaveBasis(model_ε; Ecut, kgrid)
-res_ε   = self_consistent_field(basis_ε, tol=tol)
-με = dipole(basis_ε, res_ε.ρ)
-
-#-
-polarizability = (με - μref) / ε
-
-println("Reference dipole:  $μref")
-println("Displaced dipole:  $με")
-println("Polarizability :   $polarizability")
-
-# The result on more converged grids is very close to published results.
-# For example [DOI 10.1039/C8CP03569E](https://doi.org/10.1039/C8CP03569E)
-# quotes **1.65** with LSDA and **1.38** with CCSD(T).
-
-# ## Using linear response
-# Now we use linear response to compute this analytically; we refer to standard
-# textbooks for the formalism. In the following, ``\chi_0`` is the
-# independent-particle polarizability, and ``K`` the
-# Hartree-exchange-correlation kernel. We denote with ``\delta V_{\rm ext}`` an external
-# perturbing potential (like in this case the uniform electric field). Then:
-# ```math
-# \delta\rho = \chi_0 \delta V = \chi_0 (\delta V_{\rm ext} + K \delta\rho),
-# ```
-# which implies
-# ```math
-# \delta\rho = (1-\chi_0 K)^{-1} \chi_0 \delta V_{\rm ext}.
-# ```
-# From this we identify the polarizability operator to be ``\chi = (1-\chi_0 K)^{-1} \chi_0``.
-# Numerically, we apply ``\chi`` to ``\delta V = -x`` by solving a linear equation
-# (the Dyson equation) iteratively.
-
-using KrylovKit
-
-## Apply (1- χ0 K)
-function dielectric_operator(δρ)
-    δV = apply_kernel(basis, δρ; ρ=res.ρ)
-    χ0δV = apply_χ0(res.ham, res.ψ, res.εF, res.eigenvalues, δV)
-    δρ - χ0δV
 end
 
-## δVext is the potential from a uniform field interacting with the dielectric dipole
-## of the density.
-δVext = [-a * (r[1] - 1/2) for r in r_vectors(basis)]
-δVext = cat(δVext; dims=4)
+## Function to compute the dipole for a given field strength
+function compute_dipole(ε; tol=1e-8, kwargs...)
+    scfres = self_consistent_field(make_basis_model(ε; kwargs...), tol=tol)
+    #scfres = self_consistent_field(make_basis(ε; kwargs...), tol=tol)
+    dipole(scfres.basis, scfres.ρ)
+end;
 
-## Apply χ0 once to get non-interacting dipole
-δρ_nointeract = apply_χ0(res.ham, res.ψ, res.εF, res.eigenvalues, δVext)
+# With this in place we can compute the polarizability from finite differences
+# (just like in the previous example):
+fd = let
+    ε = 0.01
+    (compute_dipole(ε) - compute_dipole(0.0)) / ε
+end
 
-## Solve Dyson equation to get interacting dipole
-δρ = linsolve(dielectric_operator, δρ_nointeract, verbosity=3)[1]
-
-println("Non-interacting polarizability: $(dipole(basis, δρ_nointeract))")
-println("Interacting polarizability:     $(dipole(basis, δρ))")
-
-# As expected, the interacting polarizability matches the finite difference
-# result. The non-interacting polarizability is higher.
+println("Polarizability (FD): ", fd)
